Update README.md
Theories in Trigonometry may be used to explain Black Holes in space as asymptotes in 3 Dimensional fields of energy or motion.

In the universe as we know it, we observe one constant motion in all observations. This constant is the appearance of trigonometric motion in the rotation of planets, stars and universes. In the milky way, we observe a rotational pattern with a high density at it's center, similar to many common vortex patterns found in nature.

If the expanse of the universe is viewed as a 3 graph, we would use trigonometric equations to graph the motion of planets and universal particles. In this research, we relate black holes with asymptotes of these trgonometric functions, and explore black holes as infinitely small ratios of space, along with the common understanding of black holes as ultra high density particles or fields in space.